refactor(crawler): crawler 상수 co-location
